refactor:move UT to seata-sqlparser-druid and arrange
