+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/dm/DmDeleteRecognizerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.rm.datasource.sql.druid.oracle;
+package org.apache.seata.sqlparser.druid.dm;
 
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
@@ -22,7 +22,6 @@ import com.alibaba.druid.sql.ast.statement.SQLDeleteStatement;
 import com.alibaba.druid.sql.dialect.oracle.ast.expr.OracleArgumentExpr;
 import org.apache.seata.sqlparser.ParametersHolder;
 import org.apache.seata.sqlparser.SQLType;
-import org.apache.seata.sqlparser.druid.oracle.OracleDeleteRecognizer;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -31,16 +30,16 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 
-public class OracleDeleteRecognizerTest {
+public class DmDeleteRecognizerTest {
 
-    private static final String DB_TYPE = "oracle";
+    private static final String DB_TYPE = "dm";
 
     @Test
     public void testGetSqlType() {
         String sql = "delete from t where id = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleDeleteRecognizer recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        DmDeleteRecognizer recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getSQLType(), SQLType.DELETE);
     }
 
@@ -49,7 +48,7 @@ public class OracleDeleteRecognizerTest {
         String sql = "delete from t where id = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleDeleteRecognizer recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        DmDeleteRecognizer recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         Assertions.assertNull(recognizer.getTableAlias());
     }
 
@@ -58,7 +57,7 @@ public class OracleDeleteRecognizerTest {
         String sql = "delete from t where id = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleDeleteRecognizer recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        DmDeleteRecognizer recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getTableName(), "t");
     }
 
@@ -67,10 +66,10 @@ public class OracleDeleteRecognizerTest {
         String sql = "delete from t";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleDeleteRecognizer recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        DmDeleteRecognizer recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition(new ParametersHolder() {
             @Override
-            public Map<Integer,ArrayList<Object>> getParameters() {
+            public Map<Integer, ArrayList<Object>> getParameters() {
                 return null;
             }
         }, new ArrayList<>());
@@ -81,10 +80,10 @@ public class OracleDeleteRecognizerTest {
         sql = "delete from t where id = ?";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         whereCondition = recognizer.getWhereCondition(new ParametersHolder() {
             @Override
-            public Map<Integer,ArrayList<Object>> getParameters() {
+            public Map<Integer, ArrayList<Object>> getParameters() {
                 ArrayList<Object> idParam = new ArrayList<>();
                 idParam.add(1);
                 Map result = new HashMap();
@@ -98,10 +97,10 @@ public class OracleDeleteRecognizerTest {
 
         sql = "delete from t where id in (?)";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         whereCondition = recognizer.getWhereCondition(new ParametersHolder() {
             @Override
-            public Map<Integer,ArrayList<Object>> getParameters() {
+            public Map<Integer, ArrayList<Object>> getParameters() {
                 ArrayList<Object> idParam = new ArrayList<>();
                 idParam.add(1);
                 Map result = new HashMap();
@@ -115,10 +114,10 @@ public class OracleDeleteRecognizerTest {
 
         sql = "delete from t where id between ? and ?";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         whereCondition = recognizer.getWhereCondition(new ParametersHolder() {
             @Override
-            public Map<Integer,ArrayList<Object>> getParameters() {
+            public Map<Integer, ArrayList<Object>> getParameters() {
                 ArrayList<Object> idParam = new ArrayList<>();
                 idParam.add(1);
                 ArrayList<Object> idParam2 = new ArrayList<>();
@@ -138,9 +137,9 @@ public class OracleDeleteRecognizerTest {
             List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, DB_TYPE);
             SQLDeleteStatement deleteAst = (SQLDeleteStatement) sqlStatements.get(0);
             deleteAst.setWhere(new OracleArgumentExpr());
-            new OracleDeleteRecognizer(s, deleteAst).getWhereCondition(new ParametersHolder() {
+            new DmDeleteRecognizer(s, deleteAst).getWhereCondition(new ParametersHolder() {
                 @Override
-                public Map<Integer,ArrayList<Object>> getParameters() {
+                public Map<Integer, ArrayList<Object>> getParameters() {
                     return new HashMap<>();
                 }
             }, new ArrayList<>());
@@ -153,7 +152,7 @@ public class OracleDeleteRecognizerTest {
         String sql = "delete from t";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleDeleteRecognizer recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        DmDeleteRecognizer recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition();
 
         //test for no condition
@@ -162,7 +161,7 @@ public class OracleDeleteRecognizerTest {
         sql = "delete from t where id = 1";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         whereCondition = recognizer.getWhereCondition();
 
         //test for normal sql
@@ -170,7 +169,7 @@ public class OracleDeleteRecognizerTest {
 
         sql = "delete from t where id in (1)";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         whereCondition = recognizer.getWhereCondition();
 
         //test for sql with in
@@ -178,7 +177,7 @@ public class OracleDeleteRecognizerTest {
 
         sql = "delete from t where id between 1 and 2";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new OracleDeleteRecognizer(sql, asts.get(0));
+        recognizer = new DmDeleteRecognizer(sql, asts.get(0));
         whereCondition = recognizer.getWhereCondition();
         //test for sql with in
         Assertions.assertEquals("id BETWEEN 1 AND 2", whereCondition);
@@ -189,7 +188,7 @@ public class OracleDeleteRecognizerTest {
             List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, DB_TYPE);
             SQLDeleteStatement deleteAst = (SQLDeleteStatement) sqlStatements.get(0);
             deleteAst.setWhere(new OracleArgumentExpr());
-            new OracleDeleteRecognizer(s, deleteAst).getWhereCondition();
+            new DmDeleteRecognizer(s, deleteAst).getWhereCondition();
         });
     }
 }
