+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/kingbase/KingbaseSelectForUpdateRecognizerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.rm.datasource.sql.druid.oscar;
+package org.apache.seata.sqlparser.druid.kingbase;
 
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
@@ -22,7 +22,7 @@ import com.alibaba.druid.sql.ast.statement.SQLSelectStatement;
 import org.apache.seata.sqlparser.ParametersHolder;
 import org.apache.seata.sqlparser.SQLParsingException;
 import org.apache.seata.sqlparser.SQLType;
-import org.apache.seata.sqlparser.druid.oscar.OscarSelectForUpdateRecognizer;
+import org.apache.seata.sqlparser.druid.kingbase.KingbaseSelectForUpdateRecognizer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -30,19 +30,17 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-/**
- * The type Oscar select for update recognizer test.
- */
-public class OscarSelectForUpdateRecognizerTest {
 
-    private static final String DB_TYPE = "oscar";
+public class KingbaseSelectForUpdateRecognizerTest {
+
+    private static final String DB_TYPE = "kingbase";
 
     @Test
     public void testGetSqlType() {
         String sql = "select * from t where id = ? for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
+        KingbaseSelectForUpdateRecognizer recognizer = new KingbaseSelectForUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getSQLType(), SQLType.SELECT_FOR_UPDATE);
     }
 
@@ -52,7 +50,7 @@ public class OscarSelectForUpdateRecognizerTest {
         String sql = "select * from t for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
+        KingbaseSelectForUpdateRecognizer recognizer = new KingbaseSelectForUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition(new ParametersHolder() {
             @Override
             public Map<Integer,ArrayList<Object>> getParameters() {
@@ -67,7 +65,7 @@ public class OscarSelectForUpdateRecognizerTest {
         String sql = "select * from t for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
+        KingbaseSelectForUpdateRecognizer recognizer = new KingbaseSelectForUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition();
 
         Assertions.assertEquals("", whereCondition);
@@ -78,7 +76,7 @@ public class OscarSelectForUpdateRecognizerTest {
             List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, DB_TYPE);
             SQLSelectStatement selectAst = (SQLSelectStatement) sqlStatements.get(0);
             selectAst.setSelect(null);
-            new OscarSelectForUpdateRecognizer(s, selectAst).getWhereCondition();
+            new KingbaseSelectForUpdateRecognizer(s, selectAst).getWhereCondition();
         });
 
         //test for query was null
@@ -87,7 +85,7 @@ public class OscarSelectForUpdateRecognizerTest {
             List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, DB_TYPE);
             SQLSelectStatement selectAst = (SQLSelectStatement) sqlStatements.get(0);
             selectAst.getSelect().setQuery(null);
-            new OscarSelectForUpdateRecognizer(s, selectAst).getWhereCondition();
+            new KingbaseSelectForUpdateRecognizer(s, selectAst).getWhereCondition();
         });
     }
 
@@ -96,7 +94,7 @@ public class OscarSelectForUpdateRecognizerTest {
         String sql = "select * from t where id = ? for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
+        KingbaseSelectForUpdateRecognizer recognizer = new KingbaseSelectForUpdateRecognizer(sql, asts.get(0));
         Assertions.assertNull(recognizer.getTableAlias());
     }
 
@@ -105,7 +103,7 @@ public class OscarSelectForUpdateRecognizerTest {
         String sql = "select * from t where id = ? for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
+        KingbaseSelectForUpdateRecognizer recognizer = new KingbaseSelectForUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getTableName(), "t");
     }
 }
