+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/kingbase/KingbaseUpdateRecognizerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.rm.datasource.sql.druid.oscar;
+package org.apache.seata.sqlparser.druid.kingbase;
 
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
@@ -24,7 +24,7 @@ import com.alibaba.druid.sql.dialect.oracle.ast.expr.OracleCursorExpr;
 import org.apache.seata.sqlparser.ParametersHolder;
 import org.apache.seata.sqlparser.SQLParsingException;
 import org.apache.seata.sqlparser.SQLType;
-import org.apache.seata.sqlparser.druid.oscar.OscarUpdateRecognizer;
+import org.apache.seata.sqlparser.druid.kingbase.KingbaseUpdateRecognizer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -32,19 +32,17 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-/**
- * The Oscar Update Recognizer Test.
- */
-public class OscarUpdateRecognizerTest {
 
-    private static final String DB_TYPE = "oscar";
+public class KingbaseUpdateRecognizerTest {
+
+    private static final String DB_TYPE = "kingbase";
 
     @Test
     public void testGetSqlType() {
         String sql = "update t set n = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
+        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getSQLType(), SQLType.UPDATE);
     }
 
@@ -53,14 +51,14 @@ public class OscarUpdateRecognizerTest {
         // test with normal
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
+        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
         List<String> updateColumns = recognizer.getUpdateColumns();
         Assertions.assertEquals(updateColumns.size(), 3);
 
         // test with alias
         sql = "update t set a.a = ?, a.b = ?, a.c = ?";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
+        recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
         updateColumns = recognizer.getUpdateColumns();
         Assertions.assertEquals(updateColumns.size(), 3);
 
@@ -73,8 +71,8 @@ public class OscarUpdateRecognizerTest {
             for (SQLUpdateSetItem updateSetItem : updateSetItems) {
                 updateSetItem.setColumn(new OracleCursorExpr());
             }
-            OscarUpdateRecognizer oscarUpdateRecognizer = new OscarUpdateRecognizer(s, sqlUpdateStatement);
-            oscarUpdateRecognizer.getUpdateColumns();
+            KingbaseUpdateRecognizer kingbaseUpdateRecognizer = new KingbaseUpdateRecognizer(s, sqlUpdateStatement);
+            kingbaseUpdateRecognizer.getUpdateColumns();
         });
     }
 
@@ -83,14 +81,14 @@ public class OscarUpdateRecognizerTest {
         // test with normal
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
+        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
         List<Object> updateValues = recognizer.getUpdateValues();
         Assertions.assertEquals(updateValues.size(), 3);
 
         // test with values
         sql = "update t set a = 1, b = 2, c = 3";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
+        recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
         updateValues = recognizer.getUpdateValues();
         Assertions.assertEquals(updateValues.size(), 3);
 
@@ -103,8 +101,8 @@ public class OscarUpdateRecognizerTest {
             for (SQLUpdateSetItem updateSetItem : updateSetItems) {
                 updateSetItem.setValue(new OracleCursorExpr());
             }
-            OscarUpdateRecognizer oscarUpdateRecognizer = new OscarUpdateRecognizer(s, sqlUpdateStatement);
-            oscarUpdateRecognizer.getUpdateValues();
+            KingbaseUpdateRecognizer kingbaseUpdateRecognizer = new KingbaseUpdateRecognizer(s, sqlUpdateStatement);
+            kingbaseUpdateRecognizer.getUpdateValues();
         });
     }
 
@@ -114,7 +112,7 @@ public class OscarUpdateRecognizerTest {
         String sql = "update t set a = 1";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
+        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition(new ParametersHolder() {
             @Override
             public Map<Integer,ArrayList<Object>> getParameters() {
@@ -131,7 +129,7 @@ public class OscarUpdateRecognizerTest {
         String sql = "update t set a = 1";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
+        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition();
 
         Assertions.assertEquals("", whereCondition);
@@ -142,7 +140,7 @@ public class OscarUpdateRecognizerTest {
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
+        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
         Assertions.assertNull(recognizer.getTableAlias());
     }
 
@@ -151,7 +149,7 @@ public class OscarUpdateRecognizerTest {
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
+        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getTableName(), "t");
     }
 }
