+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/mariadb/MariadbSelectForUpdateRecognizerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.sqlparser.druid;
+package org.apache.seata.sqlparser.druid.mariadb;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -23,7 +23,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.apache.seata.sqlparser.druid.mariadb.MariadbSelectForUpdateRecognizer;
+import org.apache.seata.sqlparser.druid.AbstractRecognizerTest;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
