+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/mysql/MySQLInsertRecognizerTest.java
@@ -14,16 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.sqlparser.druid;
+package org.apache.seata.sqlparser.druid.mysql;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-
-import org.apache.seata.sqlparser.druid.mariadb.MariadbInsertRecognizer;
-import org.apache.seata.sqlparser.druid.mysql.MySQLInsertRecognizer;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
@@ -32,13 +27,16 @@ import com.alibaba.druid.sql.dialect.mysql.ast.expr.MySqlOrderingExpr;
 
 import org.apache.seata.sqlparser.SQLParsingException;
 import org.apache.seata.sqlparser.SQLType;
+import org.apache.seata.sqlparser.druid.AbstractRecognizerTest;
 import org.apache.seata.sqlparser.util.JdbcConstants;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 /**
- * The type Mariadb insert recognizer test.
+ * The type My sql insert recognizer test.
  *
  */
-public class MariadbInsertRecognizerTest extends AbstractRecognizerTest {
+public class MySQLInsertRecognizerTest extends AbstractRecognizerTest {
 
     private final int pkIndex = 0;
 
@@ -52,13 +50,13 @@ public class MariadbInsertRecognizerTest extends AbstractRecognizerTest {
 
         SQLStatement statement = getSQLStatement(sql);
 
-        MariadbInsertRecognizer insertRecognizer = new MariadbInsertRecognizer(sql, statement);
+        MySQLInsertRecognizer mySQLInsertRecognizer = new MySQLInsertRecognizer(sql, statement);
 
-        Assertions.assertEquals(sql, insertRecognizer.getOriginalSQL());
-        Assertions.assertEquals("t1", insertRecognizer.getTableName());
-        Assertions.assertEquals(Collections.singletonList("name"), insertRecognizer.getInsertColumns());
-        Assertions.assertEquals(1, insertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).size());
-        Assertions.assertEquals(Collections.singletonList("name1"), insertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(0));
+        Assertions.assertEquals(sql, mySQLInsertRecognizer.getOriginalSQL());
+        Assertions.assertEquals("t1", mySQLInsertRecognizer.getTableName());
+        Assertions.assertEquals(Collections.singletonList("name"), mySQLInsertRecognizer.getInsertColumns());
+        Assertions.assertEquals(1, mySQLInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).size());
+        Assertions.assertEquals(Collections.singletonList("name1"), mySQLInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(0));
     }
 
     /**
@@ -71,13 +69,13 @@ public class MariadbInsertRecognizerTest extends AbstractRecognizerTest {
 
         SQLStatement statement = getSQLStatement(sql);
 
-        MariadbInsertRecognizer insertRecognizer = new MariadbInsertRecognizer(sql, statement);
+        MySQLInsertRecognizer mySQLInsertRecognizer = new MySQLInsertRecognizer(sql, statement);
 
-        Assertions.assertEquals(sql, insertRecognizer.getOriginalSQL());
-        Assertions.assertEquals("t1", insertRecognizer.getTableName());
-        Assertions.assertEquals(Arrays.asList("name1", "name2"), insertRecognizer.getInsertColumns());
-        Assertions.assertEquals(1, insertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).size());
-        Assertions.assertEquals(Arrays.asList("name1", "name2"), insertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(0));
+        Assertions.assertEquals(sql, mySQLInsertRecognizer.getOriginalSQL());
+        Assertions.assertEquals("t1", mySQLInsertRecognizer.getTableName());
+        Assertions.assertEquals(Arrays.asList("name1", "name2"), mySQLInsertRecognizer.getInsertColumns());
+        Assertions.assertEquals(1, mySQLInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).size());
+        Assertions.assertEquals(Arrays.asList("name1", "name2"), mySQLInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(0));
     }
 
     /**
@@ -90,32 +88,32 @@ public class MariadbInsertRecognizerTest extends AbstractRecognizerTest {
 
         SQLStatement statement = getSQLStatement(sql);
 
-        MariadbInsertRecognizer insertRecognizer = new MariadbInsertRecognizer(sql, statement);
+        MySQLInsertRecognizer mySQLInsertRecognizer = new MySQLInsertRecognizer(sql, statement);
 
-        Assertions.assertEquals(sql, insertRecognizer.getOriginalSQL());
-        Assertions.assertEquals("t1", insertRecognizer.getTableName());
-        Assertions.assertEquals(Arrays.asList("name1", "name2"), insertRecognizer.getInsertColumns());
-        Assertions.assertEquals(3, insertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).size());
-        Assertions.assertEquals(Arrays.asList("name1", "name2"), insertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(0));
-        Assertions.assertEquals(Arrays.asList("name3", "name4"), insertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(1));
-        Assertions.assertEquals(Arrays.asList("name5", "name6"), insertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(2));
+        Assertions.assertEquals(sql, mySQLInsertRecognizer.getOriginalSQL());
+        Assertions.assertEquals("t1", mySQLInsertRecognizer.getTableName());
+        Assertions.assertEquals(Arrays.asList("name1", "name2"), mySQLInsertRecognizer.getInsertColumns());
+        Assertions.assertEquals(3, mySQLInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).size());
+        Assertions.assertEquals(Arrays.asList("name1", "name2"), mySQLInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(0));
+        Assertions.assertEquals(Arrays.asList("name3", "name4"), mySQLInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(1));
+        Assertions.assertEquals(Arrays.asList("name5", "name6"), mySQLInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex)).get(2));
     }
 
     @Test
     public void testGetSqlType() {
         String sql = "insert into t(id) values (?)";
-        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MARIADB);
+        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MYSQL);
 
-        MariadbInsertRecognizer recognizer = new MariadbInsertRecognizer(sql, asts.get(0));
+        MySQLInsertRecognizer recognizer = new MySQLInsertRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getSQLType(), SQLType.INSERT);
     }
 
     @Test
     public void testGetTableAlias() {
         String sql = "insert into t(id) values (?)";
-        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MARIADB);
+        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MYSQL);
 
-        MariadbInsertRecognizer recognizer = new MariadbInsertRecognizer(sql, asts.get(0));
+        MySQLInsertRecognizer recognizer = new MySQLInsertRecognizer(sql, asts.get(0));
         Assertions.assertNull(recognizer.getTableAlias());
     }
 
@@ -124,28 +122,28 @@ public class MariadbInsertRecognizerTest extends AbstractRecognizerTest {
 
         //test for no column
         String sql = "insert into t values (?)";
-        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MARIADB);
+        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MYSQL);
 
-        MariadbInsertRecognizer recognizer = new MariadbInsertRecognizer(sql, asts.get(0));
+        MySQLInsertRecognizer recognizer = new MySQLInsertRecognizer(sql, asts.get(0));
         List<String> insertColumns = recognizer.getInsertColumns();
         Assertions.assertNull(insertColumns);
 
         //test for normal
         sql = "insert into t(a) values (?)";
-        asts = SQLUtils.parseStatements(sql, JdbcConstants.MARIADB);
+        asts = SQLUtils.parseStatements(sql, JdbcConstants.MYSQL);
 
-        recognizer = new MariadbInsertRecognizer(sql, asts.get(0));
+        recognizer = new MySQLInsertRecognizer(sql, asts.get(0));
         insertColumns = recognizer.getInsertColumns();
         Assertions.assertEquals(1, insertColumns.size());
 
         //test for exception
         Assertions.assertThrows(SQLParsingException.class, () -> {
             String s = "insert into t(a) values (?)";
-            List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, JdbcConstants.MARIADB);
+            List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, JdbcConstants.MYSQL);
             SQLInsertStatement sqlInsertStatement = (SQLInsertStatement)sqlStatements.get(0);
             sqlInsertStatement.getColumns().add(new MySqlOrderingExpr());
 
-            MariadbInsertRecognizer oracleInsertRecognizer = new MariadbInsertRecognizer(s, sqlInsertStatement);
+            MySQLInsertRecognizer oracleInsertRecognizer = new MySQLInsertRecognizer(s, sqlInsertStatement);
             oracleInsertRecognizer.getInsertColumns();
         });
     }
@@ -154,38 +152,37 @@ public class MariadbInsertRecognizerTest extends AbstractRecognizerTest {
     public void testGetInsertRows() {
         //test for null value
         String sql = "insert into t(id, no, name, age, time) values (1, null, 'a', ?, now())";
-        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MARIADB);
+        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MYSQL);
 
-        MariadbInsertRecognizer recognizer = new MariadbInsertRecognizer(sql, asts.get(0));
+        MySQLInsertRecognizer recognizer = new MySQLInsertRecognizer(sql, asts.get(0));
         List<List<Object>> insertRows = recognizer.getInsertRows(Collections.singletonList(pkIndex));
         Assertions.assertEquals(1, insertRows.size());
 
         //test for exception
         Assertions.assertThrows(SQLParsingException.class, () -> {
             String s = "insert into t(a) values (?)";
-            List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, JdbcConstants.MARIADB);
+            List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, JdbcConstants.MYSQL);
             SQLInsertStatement sqlInsertStatement = (SQLInsertStatement)sqlStatements.get(0);
             sqlInsertStatement.getValuesList().get(0).getValues().set(pkIndex, new MySqlOrderingExpr());
 
-            MariadbInsertRecognizer insertRecognizer = new MariadbInsertRecognizer(s, sqlInsertStatement);
-            insertRecognizer.getInsertRows(Collections.singletonList(pkIndex));
+            MySQLInsertRecognizer mysqlInsertRecognizer = new MySQLInsertRecognizer(s, sqlInsertStatement);
+            mysqlInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex));
         });
     }
 
     @Override
     public String getDbType() {
-        return JdbcConstants.MARIADB;
+        return JdbcConstants.MYSQL;
     }
 
     @Test
     public void testGetInsertColumns_2() {
         String sql = "insert into t(`id`, `no`, `name`, `age`) values (1, 'no001', 'aaa', '20')";
-        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MARIADB);
-        MariadbInsertRecognizer recognizer = new MariadbInsertRecognizer(sql, asts.get(0));
+        List<SQLStatement> asts = SQLUtils.parseStatements(sql, JdbcConstants.MYSQL);
+        MySQLInsertRecognizer recognizer = new MySQLInsertRecognizer(sql, asts.get(0));
         List<String> insertColumns = recognizer.getInsertColumns();
         for (String insertColumn : insertColumns) {
             Assertions.assertTrue(insertColumn.contains("`"));
         }
     }
-
 }
