+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/oracle/OracleInsertRecognizerTest.java
@@ -14,33 +14,32 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.rm.datasource.sql.druid.kingbase;
+package org.apache.seata.sqlparser.druid.oracle;
 
+import java.util.Collections;
+import java.util.List;
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
 import com.alibaba.druid.sql.ast.statement.SQLInsertStatement;
 import com.alibaba.druid.sql.dialect.oracle.ast.expr.OracleBinaryDoubleExpr;
 import org.apache.seata.sqlparser.SQLParsingException;
 import org.apache.seata.sqlparser.SQLType;
-import org.apache.seata.sqlparser.druid.kingbase.KingbaseInsertRecognizer;
+import org.apache.seata.sqlparser.druid.oracle.OracleInsertRecognizer;
 import org.apache.seata.sqlparser.struct.NotPlaceholderExpr;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import java.util.Collections;
-import java.util.List;
 
+public class OracleInsertRecognizerTest {
 
-public class KingbaseInsertRecognizerTest {
-
-    private static final String DB_TYPE = "kingbase";
+    private static final String DB_TYPE = "oracle";
 
     @Test
     public void testGetSqlType() {
         String sql = "insert into t(id) values (?)";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseInsertRecognizer recognizer = new KingbaseInsertRecognizer(sql, asts.get(0));
+        OracleInsertRecognizer recognizer = new OracleInsertRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getSQLType(), SQLType.INSERT);
     }
 
@@ -49,7 +48,7 @@ public class KingbaseInsertRecognizerTest {
         String sql = "insert into t(id) values (?)";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseInsertRecognizer recognizer = new KingbaseInsertRecognizer(sql, asts.get(0));
+        OracleInsertRecognizer recognizer = new OracleInsertRecognizer(sql, asts.get(0));
         Assertions.assertNull(recognizer.getTableAlias());
     }
 
@@ -58,7 +57,7 @@ public class KingbaseInsertRecognizerTest {
         String sql = "insert into t(id) values (?)";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseInsertRecognizer recognizer = new KingbaseInsertRecognizer(sql, asts.get(0));
+        OracleInsertRecognizer recognizer = new OracleInsertRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getTableName(), "t");
     }
 
@@ -69,7 +68,7 @@ public class KingbaseInsertRecognizerTest {
         String sql = "insert into t values (?)";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseInsertRecognizer recognizer = new KingbaseInsertRecognizer(sql, asts.get(0));
+        OracleInsertRecognizer recognizer = new OracleInsertRecognizer(sql, asts.get(0));
         List<String> insertColumns = recognizer.getInsertColumns();
         Assertions.assertNull(insertColumns);
 
@@ -77,7 +76,7 @@ public class KingbaseInsertRecognizerTest {
         sql = "insert into t(a) values (?)";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        recognizer = new KingbaseInsertRecognizer(sql, asts.get(0));
+        recognizer = new OracleInsertRecognizer(sql, asts.get(0));
         insertColumns = recognizer.getInsertColumns();
         Assertions.assertEquals(1, insertColumns.size());
 
@@ -88,8 +87,8 @@ public class KingbaseInsertRecognizerTest {
             SQLInsertStatement sqlInsertStatement = (SQLInsertStatement)sqlStatements.get(0);
             sqlInsertStatement.getColumns().add(new OracleBinaryDoubleExpr());
 
-            KingbaseInsertRecognizer kingbaseInsertRecognizer = new KingbaseInsertRecognizer(s, sqlInsertStatement);
-            kingbaseInsertRecognizer.getInsertColumns();
+            OracleInsertRecognizer oracleInsertRecognizer = new OracleInsertRecognizer(s, sqlInsertStatement);
+            oracleInsertRecognizer.getInsertColumns();
         });
     }
 
@@ -100,7 +99,7 @@ public class KingbaseInsertRecognizerTest {
         String sql = "insert into t(id, no, name, age, time) values (id_seq.nextval, null, 'a', ?, now())";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseInsertRecognizer recognizer = new KingbaseInsertRecognizer(sql, asts.get(0));
+        OracleInsertRecognizer recognizer = new OracleInsertRecognizer(sql, asts.get(0));
         List<List<Object>> insertRows = recognizer.getInsertRows(Collections.singletonList(pkIndex));
         Assertions.assertEquals(1, insertRows.size());
 
@@ -111,8 +110,8 @@ public class KingbaseInsertRecognizerTest {
             SQLInsertStatement sqlInsertStatement = (SQLInsertStatement)sqlStatements.get(0);
             sqlInsertStatement.getValuesList().get(0).getValues().set(pkIndex, new OracleBinaryDoubleExpr());
 
-            KingbaseInsertRecognizer kingbaseInsertRecognizer = new KingbaseInsertRecognizer(s, sqlInsertStatement);
-            kingbaseInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex));
+            OracleInsertRecognizer oracleInsertRecognizer = new OracleInsertRecognizer(s, sqlInsertStatement);
+            oracleInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex));
         });
     }
 
@@ -121,8 +120,8 @@ public class KingbaseInsertRecognizerTest {
         String sql = "insert into test(create_time) values(sysdate)";
         List<SQLStatement> sqlStatements = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseInsertRecognizer kingbase = new KingbaseInsertRecognizer(sql, sqlStatements.get(0));
-        List<List<Object>> insertRows = kingbase.getInsertRows(Collections.singletonList(-1));
+        OracleInsertRecognizer oracle = new OracleInsertRecognizer(sql, sqlStatements.get(0));
+        List<List<Object>> insertRows = oracle.getInsertRows(Collections.singletonList(-1));
         Assertions.assertTrue(insertRows.get(0).get(0) instanceof NotPlaceholderExpr);
     }
 }
