+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/oracle/OracleUpdateRecognizerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.rm.datasource.sql.druid.kingbase;
+package org.apache.seata.sqlparser.druid.oracle;
 
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
@@ -24,7 +24,7 @@ import com.alibaba.druid.sql.dialect.oracle.ast.expr.OracleCursorExpr;
 import org.apache.seata.sqlparser.ParametersHolder;
 import org.apache.seata.sqlparser.SQLParsingException;
 import org.apache.seata.sqlparser.SQLType;
-import org.apache.seata.sqlparser.druid.kingbase.KingbaseUpdateRecognizer;
+import org.apache.seata.sqlparser.druid.oracle.OracleUpdateRecognizer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -33,16 +33,16 @@ import java.util.List;
 import java.util.Map;
 
 
-public class KingbaseUpdateRecognizerTest {
+public class OracleUpdateRecognizerTest {
 
-    private static final String DB_TYPE = "kingbase";
+    private static final String DB_TYPE = "oracle";
 
     @Test
     public void testGetSqlType() {
         String sql = "update t set n = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
+        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getSQLType(), SQLType.UPDATE);
     }
 
@@ -51,14 +51,14 @@ public class KingbaseUpdateRecognizerTest {
         // test with normal
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
+        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
         List<String> updateColumns = recognizer.getUpdateColumns();
         Assertions.assertEquals(updateColumns.size(), 3);
 
         // test with alias
         sql = "update t set a.a = ?, a.b = ?, a.c = ?";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
+        recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
         updateColumns = recognizer.getUpdateColumns();
         Assertions.assertEquals(updateColumns.size(), 3);
 
@@ -71,8 +71,8 @@ public class KingbaseUpdateRecognizerTest {
             for (SQLUpdateSetItem updateSetItem : updateSetItems) {
                 updateSetItem.setColumn(new OracleCursorExpr());
             }
-            KingbaseUpdateRecognizer kingbaseUpdateRecognizer = new KingbaseUpdateRecognizer(s, sqlUpdateStatement);
-            kingbaseUpdateRecognizer.getUpdateColumns();
+            OracleUpdateRecognizer oracleUpdateRecognizer = new OracleUpdateRecognizer(s, sqlUpdateStatement);
+            oracleUpdateRecognizer.getUpdateColumns();
         });
     }
 
@@ -81,14 +81,14 @@ public class KingbaseUpdateRecognizerTest {
         // test with normal
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
+        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
         List<Object> updateValues = recognizer.getUpdateValues();
         Assertions.assertEquals(updateValues.size(), 3);
 
         // test with values
         sql = "update t set a = 1, b = 2, c = 3";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
+        recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
         updateValues = recognizer.getUpdateValues();
         Assertions.assertEquals(updateValues.size(), 3);
 
@@ -101,8 +101,8 @@ public class KingbaseUpdateRecognizerTest {
             for (SQLUpdateSetItem updateSetItem : updateSetItems) {
                 updateSetItem.setValue(new OracleCursorExpr());
             }
-            KingbaseUpdateRecognizer kingbaseUpdateRecognizer = new KingbaseUpdateRecognizer(s, sqlUpdateStatement);
-            kingbaseUpdateRecognizer.getUpdateValues();
+            OracleUpdateRecognizer oracleUpdateRecognizer = new OracleUpdateRecognizer(s, sqlUpdateStatement);
+            oracleUpdateRecognizer.getUpdateValues();
         });
     }
 
@@ -112,7 +112,7 @@ public class KingbaseUpdateRecognizerTest {
         String sql = "update t set a = 1";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
+        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition(new ParametersHolder() {
             @Override
             public Map<Integer,ArrayList<Object>> getParameters() {
@@ -129,7 +129,7 @@ public class KingbaseUpdateRecognizerTest {
         String sql = "update t set a = 1";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
+        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition();
 
         Assertions.assertEquals("", whereCondition);
@@ -140,7 +140,7 @@ public class KingbaseUpdateRecognizerTest {
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
+        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
         Assertions.assertNull(recognizer.getTableAlias());
     }
 
@@ -149,7 +149,7 @@ public class KingbaseUpdateRecognizerTest {
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        KingbaseUpdateRecognizer recognizer = new KingbaseUpdateRecognizer(sql, asts.get(0));
+        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getTableName(), "t");
     }
 }
