+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/oscar/OscarInsertRecognizerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.rm.datasource.sql.druid.dm;
+package org.apache.seata.sqlparser.druid.oscar;
 
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
@@ -22,24 +22,27 @@ import com.alibaba.druid.sql.ast.statement.SQLInsertStatement;
 import com.alibaba.druid.sql.dialect.oracle.ast.expr.OracleBinaryDoubleExpr;
 import org.apache.seata.sqlparser.SQLParsingException;
 import org.apache.seata.sqlparser.SQLType;
-import org.apache.seata.sqlparser.druid.dm.DmInsertRecognizer;
+import org.apache.seata.sqlparser.druid.oscar.OscarInsertRecognizer;
 import org.apache.seata.sqlparser.struct.NotPlaceholderExpr;
-import java.util.Collections;
-import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.util.Collections;
+import java.util.List;
 
-public class DmInsertRecognizerTest {
+/**
+ * The type Oscar insert recognizer test.
+ */
+public class OscarInsertRecognizerTest {
 
-    private static final String DB_TYPE = "dm";
+    private static final String DB_TYPE = "oscar";
 
     @Test
     public void testGetSqlType() {
         String sql = "insert into t(id) values (?)";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmInsertRecognizer recognizer = new DmInsertRecognizer(sql, asts.get(0));
+        OscarInsertRecognizer recognizer = new OscarInsertRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getSQLType(), SQLType.INSERT);
     }
 
@@ -48,7 +51,7 @@ public class DmInsertRecognizerTest {
         String sql = "insert into t(id) values (?)";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmInsertRecognizer recognizer = new DmInsertRecognizer(sql, asts.get(0));
+        OscarInsertRecognizer recognizer = new OscarInsertRecognizer(sql, asts.get(0));
         Assertions.assertNull(recognizer.getTableAlias());
     }
 
@@ -57,7 +60,7 @@ public class DmInsertRecognizerTest {
         String sql = "insert into t(id) values (?)";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmInsertRecognizer recognizer = new DmInsertRecognizer(sql, asts.get(0));
+        OscarInsertRecognizer recognizer = new OscarInsertRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getTableName(), "t");
     }
 
@@ -68,7 +71,7 @@ public class DmInsertRecognizerTest {
         String sql = "insert into t values (?)";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmInsertRecognizer recognizer = new DmInsertRecognizer(sql, asts.get(0));
+        OscarInsertRecognizer recognizer = new OscarInsertRecognizer(sql, asts.get(0));
         List<String> insertColumns = recognizer.getInsertColumns();
         Assertions.assertNull(insertColumns);
 
@@ -76,7 +79,7 @@ public class DmInsertRecognizerTest {
         sql = "insert into t(a) values (?)";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        recognizer = new DmInsertRecognizer(sql, asts.get(0));
+        recognizer = new OscarInsertRecognizer(sql, asts.get(0));
         insertColumns = recognizer.getInsertColumns();
         Assertions.assertEquals(1, insertColumns.size());
 
@@ -84,11 +87,11 @@ public class DmInsertRecognizerTest {
         Assertions.assertThrows(SQLParsingException.class, () -> {
             String s = "insert into t(a) values (?)";
             List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, DB_TYPE);
-            SQLInsertStatement sqlInsertStatement = (SQLInsertStatement) sqlStatements.get(0);
+            SQLInsertStatement sqlInsertStatement = (SQLInsertStatement)sqlStatements.get(0);
             sqlInsertStatement.getColumns().add(new OracleBinaryDoubleExpr());
 
-            DmInsertRecognizer dmInsertRecognizer = new DmInsertRecognizer(s, sqlInsertStatement);
-            dmInsertRecognizer.getInsertColumns();
+            OscarInsertRecognizer oscarInsertRecognizer = new OscarInsertRecognizer(s, sqlInsertStatement);
+            oscarInsertRecognizer.getInsertColumns();
         });
     }
 
@@ -99,7 +102,7 @@ public class DmInsertRecognizerTest {
         String sql = "insert into t(id, no, name, age, time) values (id_seq.nextval, null, 'a', ?, now())";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmInsertRecognizer recognizer = new DmInsertRecognizer(sql, asts.get(0));
+        OscarInsertRecognizer recognizer = new OscarInsertRecognizer(sql, asts.get(0));
         List<List<Object>> insertRows = recognizer.getInsertRows(Collections.singletonList(pkIndex));
         Assertions.assertEquals(1, insertRows.size());
 
@@ -107,11 +110,11 @@ public class DmInsertRecognizerTest {
         Assertions.assertThrows(SQLParsingException.class, () -> {
             String s = "insert into t(a) values (?)";
             List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, DB_TYPE);
-            SQLInsertStatement sqlInsertStatement = (SQLInsertStatement) sqlStatements.get(0);
+            SQLInsertStatement sqlInsertStatement = (SQLInsertStatement)sqlStatements.get(0);
             sqlInsertStatement.getValuesList().get(0).getValues().set(pkIndex, new OracleBinaryDoubleExpr());
 
-            DmInsertRecognizer dmInsertRecognizer = new DmInsertRecognizer(s, sqlInsertStatement);
-            dmInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex));
+            OscarInsertRecognizer oscarInsertRecognizer = new OscarInsertRecognizer(s, sqlInsertStatement);
+            oscarInsertRecognizer.getInsertRows(Collections.singletonList(pkIndex));
         });
     }
 
@@ -119,10 +122,9 @@ public class DmInsertRecognizerTest {
     public void testNotPlaceholder_giveValidPkIndex() {
         String sql = "insert into test(create_time) values(sysdate)";
         List<SQLStatement> sqlStatements = SQLUtils.parseStatements(sql, DB_TYPE);
-        ;
 
-        DmInsertRecognizer dm = new DmInsertRecognizer(sql, sqlStatements.get(0));
-        List<List<Object>> insertRows = dm.getInsertRows(Collections.singletonList(-1));
+        OscarInsertRecognizer oscar = new OscarInsertRecognizer(sql, sqlStatements.get(0));
+        List<List<Object>> insertRows = oscar.getInsertRows(Collections.singletonList(-1));
         Assertions.assertTrue(insertRows.get(0).get(0) instanceof NotPlaceholderExpr);
     }
 }
