+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/oscar/OscarSelectForUpdateRecognizerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.rm.datasource.sql.druid.dm;
+package org.apache.seata.sqlparser.druid.oscar;
 
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
@@ -22,36 +22,40 @@ import com.alibaba.druid.sql.ast.statement.SQLSelectStatement;
 import org.apache.seata.sqlparser.ParametersHolder;
 import org.apache.seata.sqlparser.SQLParsingException;
 import org.apache.seata.sqlparser.SQLType;
-import org.apache.seata.sqlparser.druid.dm.DmSelectForUpdateRecognizer;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import org.apache.seata.sqlparser.druid.oscar.OscarSelectForUpdateRecognizer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
-public class DmSelectForUpdateRecognizerTest {
+/**
+ * The type Oscar select for update recognizer test.
+ */
+public class OscarSelectForUpdateRecognizerTest {
 
-    private static final String DB_TYPE = "dm";
+    private static final String DB_TYPE = "oscar";
 
     @Test
     public void testGetSqlType() {
         String sql = "select * from t where id = ? for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmSelectForUpdateRecognizer recognizer = new DmSelectForUpdateRecognizer(sql, asts.get(0));
+        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getSQLType(), SQLType.SELECT_FOR_UPDATE);
     }
+
 
     @Test
     public void testGetWhereCondition_0() {
         String sql = "select * from t for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmSelectForUpdateRecognizer recognizer = new DmSelectForUpdateRecognizer(sql, asts.get(0));
+        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition(new ParametersHolder() {
             @Override
-            public Map<Integer, ArrayList<Object>> getParameters() {
+            public Map<Integer,ArrayList<Object>> getParameters() {
                 return null;
             }
         }, new ArrayList<>());
@@ -63,7 +67,7 @@ public class DmSelectForUpdateRecognizerTest {
         String sql = "select * from t for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmSelectForUpdateRecognizer recognizer = new DmSelectForUpdateRecognizer(sql, asts.get(0));
+        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition();
 
         Assertions.assertEquals("", whereCondition);
@@ -74,7 +78,7 @@ public class DmSelectForUpdateRecognizerTest {
             List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, DB_TYPE);
             SQLSelectStatement selectAst = (SQLSelectStatement) sqlStatements.get(0);
             selectAst.setSelect(null);
-            new DmSelectForUpdateRecognizer(s, selectAst).getWhereCondition();
+            new OscarSelectForUpdateRecognizer(s, selectAst).getWhereCondition();
         });
 
         //test for query was null
@@ -83,7 +87,7 @@ public class DmSelectForUpdateRecognizerTest {
             List<SQLStatement> sqlStatements = SQLUtils.parseStatements(s, DB_TYPE);
             SQLSelectStatement selectAst = (SQLSelectStatement) sqlStatements.get(0);
             selectAst.getSelect().setQuery(null);
-            new DmSelectForUpdateRecognizer(s, selectAst).getWhereCondition();
+            new OscarSelectForUpdateRecognizer(s, selectAst).getWhereCondition();
         });
     }
 
@@ -92,7 +96,7 @@ public class DmSelectForUpdateRecognizerTest {
         String sql = "select * from t where id = ? for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmSelectForUpdateRecognizer recognizer = new DmSelectForUpdateRecognizer(sql, asts.get(0));
+        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
         Assertions.assertNull(recognizer.getTableAlias());
     }
 
@@ -101,7 +105,7 @@ public class DmSelectForUpdateRecognizerTest {
         String sql = "select * from t where id = ? for update";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        DmSelectForUpdateRecognizer recognizer = new DmSelectForUpdateRecognizer(sql, asts.get(0));
+        OscarSelectForUpdateRecognizer recognizer = new OscarSelectForUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getTableName(), "t");
     }
 }
