+++ sqlparser/seata-sqlparser-druid/src/test/java/org/apache/seata/sqlparser/druid/oscar/OscarUpdateRecognizerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.seata.rm.datasource.sql.druid.oracle;
+package org.apache.seata.sqlparser.druid.oscar;
 
 import com.alibaba.druid.sql.SQLUtils;
 import com.alibaba.druid.sql.ast.SQLStatement;
@@ -24,7 +24,7 @@ import com.alibaba.druid.sql.dialect.oracle.ast.expr.OracleCursorExpr;
 import org.apache.seata.sqlparser.ParametersHolder;
 import org.apache.seata.sqlparser.SQLParsingException;
 import org.apache.seata.sqlparser.SQLType;
-import org.apache.seata.sqlparser.druid.oracle.OracleUpdateRecognizer;
+import org.apache.seata.sqlparser.druid.oscar.OscarUpdateRecognizer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -32,17 +32,19 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+/**
+ * The Oscar Update Recognizer Test.
+ */
+public class OscarUpdateRecognizerTest {
 
-public class OracleUpdateRecognizerTest {
-
-    private static final String DB_TYPE = "oracle";
+    private static final String DB_TYPE = "oscar";
 
     @Test
     public void testGetSqlType() {
         String sql = "update t set n = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
+        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getSQLType(), SQLType.UPDATE);
     }
 
@@ -51,14 +53,14 @@ public class OracleUpdateRecognizerTest {
         // test with normal
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
+        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
         List<String> updateColumns = recognizer.getUpdateColumns();
         Assertions.assertEquals(updateColumns.size(), 3);
 
         // test with alias
         sql = "update t set a.a = ?, a.b = ?, a.c = ?";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
+        recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
         updateColumns = recognizer.getUpdateColumns();
         Assertions.assertEquals(updateColumns.size(), 3);
 
@@ -71,8 +73,8 @@ public class OracleUpdateRecognizerTest {
             for (SQLUpdateSetItem updateSetItem : updateSetItems) {
                 updateSetItem.setColumn(new OracleCursorExpr());
             }
-            OracleUpdateRecognizer oracleUpdateRecognizer = new OracleUpdateRecognizer(s, sqlUpdateStatement);
-            oracleUpdateRecognizer.getUpdateColumns();
+            OscarUpdateRecognizer oscarUpdateRecognizer = new OscarUpdateRecognizer(s, sqlUpdateStatement);
+            oscarUpdateRecognizer.getUpdateColumns();
         });
     }
 
@@ -81,14 +83,14 @@ public class OracleUpdateRecognizerTest {
         // test with normal
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
+        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
         List<Object> updateValues = recognizer.getUpdateValues();
         Assertions.assertEquals(updateValues.size(), 3);
 
         // test with values
         sql = "update t set a = 1, b = 2, c = 3";
         asts = SQLUtils.parseStatements(sql, DB_TYPE);
-        recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
+        recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
         updateValues = recognizer.getUpdateValues();
         Assertions.assertEquals(updateValues.size(), 3);
 
@@ -101,8 +103,8 @@ public class OracleUpdateRecognizerTest {
             for (SQLUpdateSetItem updateSetItem : updateSetItems) {
                 updateSetItem.setValue(new OracleCursorExpr());
             }
-            OracleUpdateRecognizer oracleUpdateRecognizer = new OracleUpdateRecognizer(s, sqlUpdateStatement);
-            oracleUpdateRecognizer.getUpdateValues();
+            OscarUpdateRecognizer oscarUpdateRecognizer = new OscarUpdateRecognizer(s, sqlUpdateStatement);
+            oscarUpdateRecognizer.getUpdateValues();
         });
     }
 
@@ -112,7 +114,7 @@ public class OracleUpdateRecognizerTest {
         String sql = "update t set a = 1";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
+        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition(new ParametersHolder() {
             @Override
             public Map<Integer,ArrayList<Object>> getParameters() {
@@ -129,7 +131,7 @@ public class OracleUpdateRecognizerTest {
         String sql = "update t set a = 1";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
+        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
         String whereCondition = recognizer.getWhereCondition();
 
         Assertions.assertEquals("", whereCondition);
@@ -140,7 +142,7 @@ public class OracleUpdateRecognizerTest {
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
+        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
         Assertions.assertNull(recognizer.getTableAlias());
     }
 
@@ -149,7 +151,7 @@ public class OracleUpdateRecognizerTest {
         String sql = "update t set a = ?, b = ?, c = ?";
         List<SQLStatement> asts = SQLUtils.parseStatements(sql, DB_TYPE);
 
-        OracleUpdateRecognizer recognizer = new OracleUpdateRecognizer(sql, asts.get(0));
+        OscarUpdateRecognizer recognizer = new OscarUpdateRecognizer(sql, asts.get(0));
         Assertions.assertEquals(recognizer.getTableName(), "t");
     }
 }
